chore(tests): move all tests to __tests__ subdirectory
