+++ src/openApi/v2/parser/__tests__/getServer.spec.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from 'vitest';
 
-import { getServer } from './getServer';
+import { getServer } from '../getServer';
 
 describe('getServer', () => {
     it('should produce correct result', () => {
