+++ src/openApi/v2/parser/__tests__/getServiceName.spec.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from 'vitest';
 
-import { getServiceName } from './getServiceName';
+import { getServiceName } from '../getServiceName';
 
 describe('getServiceName', () => {
     it('should produce correct result', () => {
